+++ cti_sensors/er_ukf_imu_modules/error_compensation.py
@@ -2,10 +2,10 @@ import numpy as np
 
 class GyroErrorCompensation():
     def __init__(self):
-        self.correctedOmega = np.array([0,0,0], dtype=np.float32)
-        self.measuredOmega = np.array([0,0,0], dtype=np.float32)
+        self.correctedOmega = np.array([0,0,0], dtype=np.float64)
+        self.measuredOmega = np.array([0,0,0], dtype=np.float64)
 
-        self.predictedOmegaError = np.array([0,0,0], dtype=np.float32)
+        self.predictedOmegaError = np.array([0,0,0], dtype=np.float64)
         self.calculate = True
 
     def setMeasuredOmega(self, measuredOmega):
@@ -19,7 +19,7 @@ class GyroErrorCompensation():
     def getCorrectedOmega(self):
         if(self.calculate == False):
             self.correctedOmega = self.measuredOmega + self.predictedOmegaError
-            self.predictedOmegaError = np.array([0,0,0], dtype=np.float32)
+            self.predictedOmegaError = np.array([0,0,0], dtype=np.float64)
