+++ cti_sensors/er_ukf_imu_modules/ukf.py
@@ -4,23 +4,23 @@ from scipy.linalg import ldl
 
 class UKF():
     def __init__(self):
-        self.measurement = np.array([0,0,0], dtype=np.float32)
+        self.measurement = np.array([0,0,0], dtype=np.float64)
 
 
-        I = np.eye(3, dtype=np.float32)
-        O = np.zeros((3,3), dtype=np.float32)
+        I = np.eye(3, dtype=np.float64)
+        O = np.zeros((3,3), dtype=np.float64)
         self.H = np.concatenate((I,O), axis=1)
 
-        self.estimateTheta = np.zeros(3,dtype=np.float32)
-        self.estimateOmega = np.zeros(3,dtype=np.float32)
+        self.estimateTheta = np.zeros(3,dtype=np.float64)
+        self.estimateOmega = np.zeros(3,dtype=np.float64)
 
-        self.measurement = np.zeros(3,dtype=np.float32)
+        self.measurement = np.zeros(3,dtype=np.float64)
 
-        self.estimateState = np.zeros(6, dtype=np.float32)
-        self.stateCovariance = np.eye(6,dtype=np.float32) *1
+        self.estimateState = np.zeros(6, dtype=np.float64)
+        self.stateCovariance = np.eye(6,dtype=np.float64) *1
 
-        self.processNoise = np.eye(6,dtype=np.float32)*(72e-3) #Q
-        self.measurementNoise = np.eye(3,dtype=np.float32)*(72e-3) #R
+        self.processNoise = np.eye(6,dtype=np.float64)*(72e-3) #Q
+        self.measurementNoise = np.eye(3,dtype=np.float64)*(72e-3) #R
 
         self.N = 6.0
         self.k = 3.0-self.N
@@ -34,7 +34,7 @@ class UKF():
             Calcula a matriz Wb descrita na equação (6) de [3], que possibilita computar dtheta_dt
         '''
 
-        wb = np.eye(3, dtype=np.float32)
+        wb = np.eye(3, dtype=np.float64)
 
         wb[0][1] = sin(x[0])*tan(x[1])
         wb[0][2] = cos(x[0])*tan(x[1])
@@ -58,7 +58,7 @@ class UKF():
         wY = omega[1]
         wZ = omega[2]
 
-        vB = np.zeros((3,3), dtype=np.float32)
+        vB = np.zeros((3,3), dtype=np.float64)
 
         vB[0][0] = ((cos(psi)*sin(theta)*wY)/cos(theta)) - ((sin(psi)*sin(theta)*wZ)/cos(theta))
         vB[0][1] = ((sin(psi)*wY)/np.power(cos(theta), 2)) + ((cos(psi)*wZ)/np.power(cos(theta), 2))
@@ -109,7 +109,7 @@ class UKF():
 
         L = np.linalg.cholesky(covariance)
 
-        sigmaPoint = np.zeros((13,6), dtype=np.float32)
+        sigmaPoint = np.zeros((13,6), dtype=np.float64)
 
         sigmaPoint[0] = state
 
@@ -130,15 +130,15 @@ class UKF():
         sigmaPoint = self.computeSigmaPoint(previousState, previousCovariance)
         
         #Propaga os sigma points
-        predictedStateSigma = np.zeros((13,6), dtype=np.float32)
+        predictedStateSigma = np.zeros((13,6), dtype=np.float64)
 
         for i in range(13):
             predictedStateSigma[i] = self.f(sigmaPoint[i],sigmaPoint[i][3:6],deltaT)
 
         #Calcula a predicao da media e covariancia
         
-        predictedState = np.zeros(6,dtype=np.float32)
-        predictedCovariance = np.zeros((6,6),dtype=np.float32)
+        predictedState = np.zeros(6,dtype=np.float64)
+        predictedCovariance = np.zeros((6,6),dtype=np.float64)
         
         for i in range(13):
             alpha = self.computeAlpha(i)
@@ -161,15 +161,15 @@ class UKF():
         ##Correcao##
 
         #Propagar sigmaPoint para a medicao
-        predictedMeasureSigma = np.zeros((13,3), dtype=np.float32)
+        predictedMeasureSigma = np.zeros((13,3), dtype=np.float64)
 
         for i in range(13):
             predictedMeasureSigma[i] = self.h(predictedStateSigma[i])
 
         #Estimar media e covariancia das prediçoes das medicoes
 
-        predictedMeasure = np.zeros(3,dtype = np.float32)
-        predictedMeasureCov = np.zeros((3,3), dtype=np.float32)
+        predictedMeasure = np.zeros(3,dtype = np.float64)
+        predictedMeasureCov = np.zeros((3,3), dtype=np.float64)
 
         for i in range(13):
             alpha = self.computeAlpha(i)
@@ -190,7 +190,7 @@ class UKF():
         predictedMeasureCov += self.measurementNoise
 
         #Calcular cross-covariance e Kalman gain
-        crossCovariance = np.zeros((6,3), dtype=np.float32)
+        crossCovariance = np.zeros((6,3), dtype=np.float64)
 
         for i in range(13):
             alpha = self.computeAlpha(i)
